Redirect master documentation to v1.x
This documentation is technically for the v0.x branch, but people visiting these links now will probably look for the latest version.  The master documentation is on the GitHub.io domain.

Uses a meta redirect, and a canonical link to hopefully not break SEO results.

Script used to do this, in case we need to make changes:

```php
<?php

$files = glob(__DIR__ . '/*.html');

$template = <<<'HTML'
<!DOCTYPE html>
<html lang="en">
<head>
    <meta charset="UTF-8">
    <meta http-equiv="refresh" content="0;url=/docs/1.x/$page" />
    <link rel="canonical" href="https://hydephp.com/docs/1.x/$page">
    <title>Redirecting...</title>
</head>

HTML;

foreach ($files as $file) {
    $page = basename($file, '.html');
    $content = str_replace('$page', $page, $template);
    file_put_contents($file, $content);
}
```